+++ packages/openneuro-app/src/scripts/index.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect } from "react"
-import { useLocation, useNavigate } from "react-router-dom"
+import React from "react"
 import Uploader from "./uploader/uploader"
 import AppRoutes from "./routes"
 import HeaderContainer from "./common/containers/header"
@@ -7,37 +6,12 @@ import FooterContainer from "./common/containers/footer"
 import { SearchParamsProvider } from "./search/search-params-ctx"
 import { UserModalOpenProvider } from "./utils/user-login-modal-ctx"
 import { useAnalytics } from "./utils/analytics"
-import { useUser } from "./queries/user"
 import { NotificationsProvider } from "./users/notifications/user-notifications-context"
+import { GoogleOrcidRedirect } from "./authentication/google-redirect"
 import "../assets/email-header.png"
 
 const Index: React.FC = () => {
   useAnalytics()
-
-  const navigate = useNavigate()
-  const location = useLocation()
-  const { user, loading, error } = useUser()
-
-  useEffect(() => {
-    if (
-      !loading &&
-      !error &&
-      location.pathname !== "/orcid-link" &&
-      user?.provider === "google"
-    ) {
-      navigate("/orcid-link")
-    }
-  }, [location.pathname, user, loading, error, navigate])
-
-  if (loading || error) return null
-
-  const initialNotifications = user?.notifications?.map((n) => ({
-    id: n.id,
-    status: Array.isArray(n.notificationStatus)
-      ? n.notificationStatus.map((ns) => ns.status.toLowerCase())[0] || "unread"
-      : n.notificationStatus?.status?.toLowerCase() || "unread",
-    ...n,
-  })) || []
 
   return (
     <Uploader>
@@ -48,6 +22,7 @@ const Index: React.FC = () => {
               <HeaderContainer />
               <AppRoutes />
             </div>
+            <GoogleOrcidRedirect />
           </NotificationsProvider>
           <div className="sticky-footer">
             <FooterContainer />
